pipeline/requirements.txt
psycopg2-binary==2.9.9
streamlit>=1.28.0
aiohttp==3.9.1
